VK/requirements.txt
vkbottle
DateTime
uuid
pysqlite3
pymorphy2
pytrovich
comtypes
python-pptx
ujson
cchardet
aiodns
aiohttp[speedups]
aiofiles